add dump_last_paste command
which prints the internal name_pairs buffer that holds the old and new
names for the last paste operation. this is to support more complicated
use cases beyond what mark_last_paste has to offer, such as to construct
commands to move those files/folders back to where they were.