fix local setting ordering and unintended regex match.
local setting is great, but there is a bug in its dealing with parent
and child paths. since it uses regex to match a local path, it is
inevitable that the same local setting in a parent path also matches for
a child path, because the parent path is a proper prefix of the child
path. depending on the order in iteration, when looking for a child
local setting, if the parent shows up first, then it will pick the
parent local setting, effectively rendering the child local setting
non-existent. the way to fix this is to give the child local setting
preference, and the fix is to use sorted dict and traverse it in
reverse order, so that a child path always shows up before its parent,
since by definition, a child path is its parent plus some suffix, they
will always show up first in a reverse sorted ordering.

some other possible fixes considered by abandoned:
1. make regex to match the whole string with ^$, instead of anywhere.
   this will make sure the child path won't match the parent strict
   pattern. but it also makes it impossible to have flexible regex to
   capture multiple local paths in one go.
2. use ^$ pattern in setlocal path= command. this didn't work because
   internally the path is escaped, so the ^$ get escaped into \\^\\$
   and lose their regex effect.
3. in case local path ends up taking the default fm.thisdir value, take
   it as ^fm.thisdir$ instead. this only takes care of this one special
   case and isn't general enough to accommodate other similar usages.