filter stack adds temp stack filter to allow filter as you type.
as you type works great for the filter command, and I would like to have
it to filter stack as well. turns out ranger constructs a new command
object for each char typed, so there is no way to capture the first key
for a filter_stack add command, consequently, I have to draw back to a
dedicated temporary_stack_filter data member, similar to how
temporary_filter does the work for the filter command. once this is
added to each directory, the filter stack command may change this temp
filter as user types.