local settings prefer sorted dict but fall back to dict
this is to work around envs where sorted dict isn't available and cannot
be installed either. the current impl has the drawback of not warning
users a regular dict is used instead of a sorted dict, in cases when
sorted dict is not available but it is possible to install it.