disable image orientation in preview
the identify command that is used to determine if the image exif
contains an orientation component has to create short lived temp file
under the image directory. this results in the directory mtime updated
all the time, with only a few tiny milliseconds different. ranger then
has to reload the directory because the cache time stamp is a little
behind when identify just wrote the temp file, even though the file
itself may have been removed already, and there is no meaningful change
in the directory.

once the directory is reloaded, the cursor is placed on the same image,
which, however, has to be previewed again, resulting in an infinite
loop, since the preview script will go through the same identify command
again, which will make another temp file and update the directory's mtime.

this looks like a new behavior from identify under ubuntu 20.04. I am
guessing the orientation isn't that important, since I don't think all
image viewers respect it anyway. turn it off to avoid the intricate bug.