Revert "disable image orientation in preview"
This reverts commit 9f81962366aabf32115a834a3d8d0d7176f72a84.

googled a bit why imagemagick identify has to create temp files. turns
out this is an imagemagick issue, not necessarily about identify.

basically im has some memory constraints and when it's out of mem, it
then makes temp file to compensate. more details see
https://stackoverflow.com/questions/39210038/how-can-i-configure-imagemagick-to-not-use-any-temporary-files

so one solution is to give it more memory to use. config is in
/etc/ImageMagick-6/policy.xml

change the default 256Mb into, e.g., 1024Mb
<!-- DEFAULT: <policy domain="resource" name="memory" value="256MiB"/> -->
<policy domain="resource" name="memory" value="1024MiB"/>

at least with my machine, this is good enough to avoid imagemagick to
create temp files and resolve the delicate infinite loop condition.