Revert "Revert "disable image orientation in preview""
This reverts commit 99bfb35c32288c82b6e1aec97f528312780f9358.

disable scope.sh image preview orientation detection and auto orient
feature. this previous resulted in an buggy infinite loop, which I fixed
by giving imagicmagick more memory allowance.

however, turns out it is also a perf bottleneck, in that for identify to
detect the orientation in exif, it makes the preview considerably
slower, so much so that I can easily tell with bare eyes, and at least
per my test, the orientation just says 1, which means no need for
rotation, which wastes the cycles to detect it in the first place.
again, even when there is orientation, I don't see a good use case to
rotate the image in preview, as I could just rotate the image to orient
it to my like. so this auto orient feature really doesn't add much
value, but either brings in bugs or perf hit, so turn it off again.