ranger image preview break apart preview cache logic.
this is a logically equivalent transformation, and didn't change any
semantics, but it does make debugging easier by separating out the logic
that compares the source file mtime vs. the cache image mtime to decide
if the cached image is useable. one tricky case with this approach is
that some source files have wrong mtime marked in the future for
whatever reasons, in which case the cache image mtime is always smaller
than the source file mtime because the cache image is just created, so
pretty much have current time, whereas the source file, due to whatever
reasons, have a future mtime, so the cache image just created cannot be
used, which triggers ranger to create another cache image, but won't be
useable due to the same reason, resulting in an infinite loop, as if the
source file cannot be previewed.